pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-ai-thread-bot"
version = "0.1.0"
description = "Telegram bot with thread-aware AI assistance via OpenRouter."
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "aiogram>=3.13.0,<4.0",
    "SQLAlchemy[asyncio]>=2.0.30",
    "alembic>=1.13.2",
    "asyncpg>=0.29.0",
    "openai>=1.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "PyYAML>=6.0.1",
    "structlog>=24.4.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "types-PyYAML>=6.0.12",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "ASYNC", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
